refactor: extract CLI wrapper to scripts/libs/CLI.ts
